+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Show Food Filter in Settings [#5490](https://github.com/rokwire/illinois-app/issues/5490).
 - Assistant: increase request timeout to 120 secs for the /query api [#5493](https://github.com/rokwire/illinois-app/issues/5493).
+### Removed
+- Maps Accessibility resize workaround is removed [#5051](https://github.com/rokwire/illinois-app/issues/5051)
 
 ## [7.3.17] - 2025-10-22
 ### Changed